Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src/script-src-1_1-expected.txt

CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
Inline script should not run without 'unsafe-inline' script-src directive, even for script-src 'self'.


Harness Error (TIMEOUT), message = null

PASS Inline script block
PASS Inline event handler
NOTRUN Should fire policy violation events
